Moves observer package to submodel
