+++ src/main/java/org/eclipse/basyx/aas/registration/observing/ObservableAASRegistryService.java
@@ -15,9 +15,9 @@ import java.util.List;
 
 import org.eclipse.basyx.aas.metamodel.map.descriptor.AASDescriptor;
 import org.eclipse.basyx.aas.metamodel.map.descriptor.SubmodelDescriptor;
-import org.eclipse.basyx.aas.observer.Observable;
 import org.eclipse.basyx.aas.registration.api.IAASRegistry;
 import org.eclipse.basyx.submodel.metamodel.api.identifier.IIdentifier;
+import org.eclipse.basyx.submodel.observer.Observable;
 import org.eclipse.basyx.vab.exception.provider.ProviderException;
 
 /**
