+++ src/main/java/org/eclipse/basyx/submodel/restapi/observing/ISubmodelAPIObserver.java
@@ -9,7 +9,7 @@
 ******************************************************************************/
 package org.eclipse.basyx.submodel.restapi.observing;
 
-import org.eclipse.basyx.aas.observer.IObserver;
+import org.eclipse.basyx.submodel.observer.IObserver;
 
 /**
  * Interface for an observer of {@link ObservableSubmodelAPI}
